Add internal gene references
Indels are rather rare in the internal genes (with the interesting
exception of NS), so the genes are covered well with a single reference.
I chose the H1N1 A/Puerto_Rico/8/1934 (EPI_ISL_146807) for this role
since it 1) has a complete genome sequence with no ambiguous residues
and 2) has been the subject of extensive structural research.

I retrieved all available public WGS data including around 70000 to
80000 sequences for each internal gene. For each segment (other than NS)
the 1st, 2nd and 3rd quantiles were all equal (indicating at least 50%
of the strains in each group are of exactly the same length). The
lengths of the translated A/Puerto_Rico/8/1934 segments exactly matched
these quantiles. For NS, the translated product matched the 2nd quantile
value (the median).

So overall, while A/Puerto_Rico/8/1934 may be very different in sequence
from modern strains, it will work fine as an index reference.