Do not create .tmp or .tmp_aln files
Now flutile creates proper temporary files. This is certainly not
optimal. Alignment is expensive, it would be better if the aligned files
were cached.

An easy way to do this would be to write the alignments to /var/tmp/
with the input fastafile md5sum (made by smof) as a suffix. Before
running an alignment, the md5sum of the sequence would be checked
against all files in the /var/tmp directory, if any of them match, then
that alignment is used. If we are feeling paranoid, we could clean the
alignment, take its md5sum, and compare it to the input md5sum.

However, this will lead to accumulation of rather large files in
/var/tmp/. I could add a `clear cache` argument. Meh, I'll add this to
the "that would be nice but I'm not going to do it yet" pile.