Remove mixed sequences from test M fasta
The local and remote MAFFT results were slightly different with regards
to where the placed gaps in a crazy M alignment that included a PB1.
This is why the CI on GitHub actions was failing.

So I removed the mixed strains from the M.fna file and now everything
should be fine (fingers crossed).